inline extended-time element into relative-time
