+++ src/local-time-element.ts
@@ -5,11 +5,11 @@ import RelativeTimeElement from './relative-time-element.js'
 const formatters = new WeakMap<Element, ReturnType<typeof makeFormatter>>()
 
 export default class LocalTimeElement extends RelativeTimeElement {
-  attributeChangedCallback(attrName: string, oldValue: string, newValue: string): void {
+  attributeChangedCallback(attrName: string): void {
     if (attrName === 'hour' || attrName === 'minute' || attrName === 'second' || attrName === 'time-zone-name') {
       formatters.delete(this)
     }
-    super.attributeChangedCallback(attrName, oldValue, newValue)
+    super.attributeChangedCallback(attrName)
   }
 
   // Formats the element's date, in the user's current locale, according to
@@ -28,9 +28,9 @@ export default class LocalTimeElement extends RelativeTimeElement {
   //   second  - "numeric", "2-digit"
   //
   // Returns a formatted time String.
-  get format(): string {
+  getFormattedDate(): string | undefined {
     const d = this.date
-    if (!d) return ''
+    if (!d) return
 
     const date = formatDate(this, d) || ''
     const time = formatTime(this, d) || ''
